GridFS: get rid of upload limitation
When using play 2.4 a general multipart maximum entity size is enforced
even when using a custom BodyParser for the file parts.

When trying to POST a file which exceeds the configurable limit, the
server responds with a 413 "Request Entity Too Large" status because the
maxLength value given to parse.multipartFormData(f, maxLength) is
applied to the whole body of the request, not to any specific part
thereof.

Use the Multipart.multipartParser function directly which applies the
limit only to data parts not handled by the given file part handler.