Refactor in functions from ReactiveMongo/Play-ReactiveMongo
GridFSController contains functions from the MongoController of the forked repo. I am unsure at this stage is these will remain in the current form so please consider these to be experimental

Additionally the Formatters object has been merged in but again not 100% sure we require these bindings. Leaving in for now